Return empty list if job is null in getRepoNames
DockerImageExtractor.getDockerImagesUsedByJobFromAll expects a non-null
job argument but the test environment on the master branch and on the
test branch were both passing a null job.  When the docker-workflow
plugin is included in the test dependencies, then an implementation of
the abstract method is used that expects the NonNull to be honored.

It seems like this should have been detected by spotbugs, since the
argument is annotated NonNull in the called method and is being passed
a null value from this method.

https://github.com/jenkinsci/bom/issues/2439 is the issue that detected
this implementation problem.