Update the BackCompat102Test.zip build dirs to numerical format - also keep it unzipped
While removing the `RunIdMigrator`, it was noticed that `BackCompat102Test` test class was failing in BOM weekly tests.
Reported in https://github.com/jenkinsci/jenkins/pull/10517.

The reason is build directories were historically named in `yyyy-MM-dd_HH-mm-ss` format, which was changed to numerical format in 2015 Jan.
However the zip file in here was still working because `RunIdMigrator` was renaming during runtime.

As `RunIdMigrator` is getting removed we need to fix these directory names.

Also since Zip files are hard to compare -- unzipping and checking in the files with,
* build folder renamed to numerical format
* <timestamp> tag is put in `build.xml` instead of `<number>`.
* update the buildIds in fingerprint files as the name changed from date-time format to numerical